Rename app directory to bgpcfgd
