+++ src/sonic-bgpcfgd/.gitignore
@@ -2,7 +2,7 @@
 build/
 dist/
 *.egg-info/
-app/*.pyc
+bgpcfgd/*.pyc
 tests/*.pyc
 tests/__pycache__/
 .idea
