+++ src/sonic-bgpcfgd/bgpcfgd/managers_allow_list.py
@@ -3,10 +3,10 @@ Implementation of "allow-list" feature
 """
 import re
 
-from app.log import log_debug, log_info, log_err, log_warn
-from app.template import TemplateFabric
-from app.manager import Manager
-from app.utils import run_command
+from bgpcfgd.log import log_debug, log_info, log_err, log_warn
+from bgpcfgd.template import TemplateFabric
+from bgpcfgd.manager import Manager
+from bgpcfgd.utils import run_command
 
 class BGPAllowListMgr(Manager):
     """ This class initialize "AllowList" settings """
