+++ src/sonic-bgpcfgd/bgpcfgd/managers_setsrc.py
@@ -1,8 +1,8 @@
 import jinja2
 
-from app.log import log_err, log_warn, log_info
-from app.manager import Manager
-from app.template import TemplateFabric
+from bgpcfgd.log import log_err, log_warn, log_info
+from bgpcfgd.manager import Manager
+from bgpcfgd.template import TemplateFabric
 
 
 class ZebraSetSrc(Manager):
