+++ src/sonic-bgpcfgd/pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = --cov=app --cov-report term
+addopts = --cov=bgpcfgd --cov-report term
